컴퓨터 클래스 분리
