+++ src/main/java/baseball/Computer.java
@@ -1,0 +1,25 @@
+package baseball;
+
+import camp.nextstep.edu.missionutils.Randoms;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Computer {
+    private static final int NUMBER_LENGTH = 3;
+
+    /**
+     * 컴퓨터의 수를 생성
+     * @return 1~9까지의 숫자 3개가 담긴 리스트
+     */
+    public static List<Integer> getRandomNumbers() {
+        List<Integer> numbers = new ArrayList<>();
+        while (numbers.size() < NUMBER_LENGTH){
+            int randomNumber = Randoms.pickNumberInRange(1, 9);
+            if(!numbers.contains(randomNumber)){
+                numbers.add(randomNumber);
+            }
+        }
+        return numbers;
+    }
+}
